cast int datatype to fix failure on windows

+++ tabular/tests/unittests/data/test_label_cleaner.py
@@ -79,12 +79,12 @@ def test_label_cleaner_binary():
 def test_label_cleaner_multiclass():
     # Given
     problem_type = MULTICLASS
-    input_labels_numpy = np.array([2, 4, 2, 2, 4, 1])
-    input_labels = pd.Series(input_labels_numpy)
+    input_labels_numpy = np.array([2, 4, 2, 2, 4, 1], dtype="int32")
+    input_labels = pd.Series(input_labels_numpy, dtype="int32")
     input_labels_category = input_labels.astype("category")
     input_labels_with_shifted_index = input_labels.copy()
     input_labels_with_shifted_index.index += 5
-    input_labels_new = np.array([3, 5, 2])
+    input_labels_new = np.array([3, 5, 2], dtype="int32")
     expected_output_labels = pd.Series([1, 2, 1, 1, 2, 0], dtype="uint8")
     expected_output_labels_new = pd.Series([np.nan, np.nan, 1])
     expected_output_labels_new_inverse = pd.Series([np.nan, np.nan, 2])
@@ -113,6 +113,9 @@ def test_label_cleaner_multiclass():
     output_labels_uncleaned_new = label_cleaner.transform_pred_uncleaned(y=input_labels_new)
     output_labels_uncleaned_new_inverse = label_cleaner.inverse_transform_pred_uncleaned(y=output_labels_uncleaned_new)
 
+    input_labels = input_labels.astype("int32")
+    output_labels_uncleaned_inverse = output_labels_uncleaned_inverse.astype("int32")
+    
     assert expected_output_labels.equals(output_labels)
     assert expected_output_labels.equals(output_labels_with_numpy)
     assert expected_output_labels.equals(output_labels_category)
